mv run_client_app to supernode.app
